fix(TTD): Refactor Long Condition
