+++ src/App.tsx
@@ -4,6 +4,7 @@ import "./index.css";
 import InputField from "./components/InputField";
 import TodoList from "./components/TodoList";
 import FilterTasks from "./components/FilterTasks";
+import FilteredTasks from "./components/FilteredTasks";
 import { Priority, Tasks } from "./Model";
 
 function App() {
@@ -12,6 +13,7 @@ function App() {
   const [filter, setFilter] = useState<string>('all');
   const [priorityFilter , setPriorityFilter] = useState<string>('all');
   const [dateFilter, setDateFilter] = useState<string>('all');
+  const [filteredTasks, setFilteredTasks] = useState<Tasks[]>([]);
 
   const clearCompletedTasks = () => {
     const activeTasks = tasks.filter((task) => !task.isCompleted);
@@ -36,38 +38,6 @@ function App() {
     }
   };
 
-  const filterDeadlineTasks = (task: Tasks) => {
-    const today = new Date();
-    const taskDate = new Date(task.deadline?.to);
-    
-    if (dateFilter === 'today') {
-      return taskDate.toDateString() === today.toDateString();
-    } else if (dateFilter === 'upcoming') {
-      return taskDate > today;
-    } else if (dateFilter === 'overdue') {
-      return taskDate < today;
-    }
-    return true;
-  } 
-  const filteredTasks = tasks.filter(task => {
-    const dateCondition = filterDeadlineTasks(task);
-
-    if (filter === 'all' && priorityFilter === 'all') {
-      return dateCondition;
-    } else if (filter === 'completed' && priorityFilter === 'all') {
-      return task.isCompleted && dateCondition;
-    } else if (filter === 'active' && priorityFilter === 'all') {
-      return !task.isCompleted && dateCondition;
-    } else if (filter === 'all' && priorityFilter !== 'all') {
-      return task.priority === priorityFilter && dateCondition;
-    } else if (filter === 'completed' && priorityFilter !== 'all') {
-      return task.isCompleted && task.priority === priorityFilter  && dateCondition;
-    } else if (filter === 'active' && priorityFilter !== 'all') {
-      return !task.isCompleted && task.priority === priorityFilter && dateCondition;
-    }
-    return false;
-  });
-  
   return (
     <div className="App">
       <h1 className="heading">Plan Your Thoughts! 😀 </h1>
@@ -81,6 +51,7 @@ function App() {
         <FilterTasks filter={filter} setFilter={setFilter} priorityFilter={priorityFilter} setPriorityFilter={setPriorityFilter} dateFilter={dateFilter} setDateFilter={setDateFilter} />
       </div>
       <InputField todo={todo} setTodo={setTodo} handleAdd={handleAdd} />
+      <FilteredTasks tasks={tasks} filter={filter} priorityFilter={priorityFilter} dateFilter={dateFilter} setFilteredTasks={setFilteredTasks} />
       <TodoList tasks={filteredTasks} setTasks={setTasks} />
     </div>
   );
